# s2: Reduce ReadFrom temporary allocations